Header fixes
