+++ client/web/package.json
@@ -3,9 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.32",
-    "@fortawesome/free-solid-svg-icons": "^5.15.1",
-    "@fortawesome/react-fontawesome": "^0.1.11",
     "@mapbox/mapbox-gl-geocoder": "^4.7.0",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
@@ -22,7 +19,6 @@
     "bulma": "^0.9.1",
     "bulma-start": "0.0.4",
     "bulma-switch": "^2.0.0",
-    "classnames": "^2.2.6",
     "connected-react-router": "^6.8.0",
     "gh-pages": "^3.1.0",
     "js-cookie": "^2.2.1",
