+++ client/web/src/apps/about/App.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
-import { BrowserRouter } from 'react-router-dom'
 import Routes from './Routes'
 import Div100vh from 'react-div-100vh'
 import Footer from 'components/main/Footer'
@@ -19,15 +18,13 @@ const useStyles = makeStyles({
 const App = () => {
   const classes = useStyles()
   return (
-    <BrowserRouter>
-      <Div100vh className={classes.root}>
-        <div className={classes.content}>
-          <Routes />
-        </div>
-        <Footer />
-        <CookieConsentBanner />
-      </Div100vh>
-    </BrowserRouter>
+    <Div100vh className={classes.root}>
+      <div className={classes.content}>
+        <Routes />
+      </div>
+      <Footer />
+      <CookieConsentBanner />
+    </Div100vh>
   )
 }
